==> analysis/01_prepping_pre2003_policing_data.Rmd <==
---
title: "Preparing Pre-2003 Policing Data"
author: "Angel Mendiola Ross"
date: "5/5/2020"
output: html_document
---

```{r}
knitr::opts_knit$set(root.dir = '/Users/angelmr/ps239t-final-project/')
```

## 1.0 Setup

```{r}

### Clear global environment
rm(list=ls())

# install.packages("readxl") # Needed this for reading in xlsx files which is how the state provides raw data
library(pacman)
pacman::p_unload(all)

pacman::p_load(
  tidyverse, #dplyr, readr, etc.
  plyr, #for editing factor levels
  magrittr, #%<>% operator
  readxl, #for excel workbooks
  tidyr #to transform from wide to long
)

```

## 1.1 Prepping 1992-2017 policing data

The first step is reading in policing data from the California Local Government Finance Almanac: <http://www.californiacityfinance.com/index.php#SPENDING>

I decided to supplement the data from the state's ByTheNumbers tool because this source has data going back to 1990s whereas the state's ByTheNumbers tool (<https://cities.bythenumbers.sco.ca.gov/#!/year/default>) only goes back to the year 2003. I would like to run my analysis back from 2000.

These data come in an Excel workbook and is saved in the Data/Raw folder.

```{r reading in data}
# Seeing what sheets are in the excel document
# excel_sheets("Police17p.xlsx")

# Bringing in Sheet 1 "ExpTotal" which has all of the expenditures data (Note: NAs are coded as "NR" in the dataset)
police <- read_excel("data/raw/Police17p.xlsx", sheet = 1, na = c("", "NR"))

```

## 1.2 Now to clean this data frame of all the Excel formatting in the original.

```{r cleaning data frame}
# the actual column names are the 7th row of the file, so I am renaming the columns with this information
police %<>%
  # rename with 7th row
  set_names(police[7, ])

# deleting the first 7 rows
police = police[-(1:7),]

# deleting na columns
police = police[,-c(29,30,32,33)]

```

## 1.3 Transforming data from wide to long and adding a calendar year variable

``` {r}
# transforming data from wide to long using gather
police_long <- gather(police, year, police_total, 3:28, factor_key=TRUE)

# create new year variable to be the calendar year (rather than FY)
police_long %<>%
  mutate(cal_year = NA) %>%
  mutate(cal_year = ifelse(year %in% c("1991-92"), 1992, cal_year))

# note I tried to run this in a single mutate command but it kept overwriting with NAs so I resorted to this
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("1992-93"), 1993, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("1993-94"), 1994, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("1994-95"), 1995, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("1995-96"), 1996, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("1996-97"), 1997, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("1997-98"), 1998, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("1998-99"), 1999, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("1999-00"), 2000, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2000-01"), 2001, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2001-02"), 2002, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2002-03"), 2003, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2003-04"), 2004, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2004-05"), 2005, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2005-06"), 2006, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2006-07"), 2007, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2007-08"), 2008, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2008-09"), 2009, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2009-10"), 2010, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2010-11"), 2011, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2011-12"), 2012, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2012-13"), 2013, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2013-14"), 2014, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2014-15"), 2015, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2015-16"), 2016, cal_year))
police_long %<>%
  mutate(cal_year = ifelse(year %in% c("2016-17"), 2017, cal_year))

# saving a service responsibility dataset
service_resp <- police_long %>%
  subset(cal_year==2017) %>%
  select("City", "SERVICE RESPONSIBILITY")

save(service_resp,
     file="data/city_service_responsibility.RData")

# reducing to policing data to columns on interest
police_long <- police_long[c(7,1,2,6)]

# renaming year column (in base r as plyr is being finicky)
names(police_long)[names(police_long)=="cal_year"] <- "year"

# saving 1992-2017 policing data
save(police_long,
     file="data/policing_data_9217.RData")

```
