==> analysis/04_interpolating_census_data.Rmd <==
---
title: "Joining and Interpolating Census Data"
author: "Angel Mendiola Ross"
date: "5/10/2020"
output: html_document
---

```{r setup}
knitr::opts_knit$set(root.dir = '/Users/angelmr/ps239t-final-project/')
```

## 1.0 Setup

```{r}

### Clear global environment
rm(list=ls())

# install.packages('zoo')
library(pacman)
pacman::p_unload(all)

pacman::p_load(
  tidyverse, #dplyr, readr, etc.
  magrittr, #%<>% operator
  tidyr, #to transform from wide to long
  ggthemes, # for pretty charts
  zoo, # for na.approx
  lubridate # for dates
)
```

## 1.1 Loading Census Data and Policing Data

``` {r}

load("data/census_data.RData")

# checking the variables
names(combined_data)

# adding variables I forgot to add earlier
combined_data %<>% 
  mutate(units_5plus = units_5to9 + units_10to19 + units_20to49 + units_50plus)


# trimming down to the variables of interest for my analysis
combined_data <- combined_data[c(1:3,14,4,7:13,15:33,36:39,42,43,109,65:68,70:75,78:104,106:108)]

# loading policing data
load("data/data_no_census.RData")

```

## 1.2 Joining Census Data and Policing Data

I have to linearly interpolate Census data from 2000 to 2009 because the ACS was only started in 2005 and the Census API only reports data from the 2009 ACS and beyond (I am using five-year estimates so the 2009 ACS includes data from 2005-2009).

In order to linearly interpolate the Census data, I need to have missing years. So I am first joining my Census data with my policing (and other independent varibles) data.

``` {r}
# to prepare for join, I am going to load a crosswalk I created with city financial data ids and census geoids
xwalk <- read_csv("data/city_fin_census_crosswalk.csv")

# renaming GEOID and dropping the first GEOID
xwalk %<>% dplyr::rename(city_name = entity_name)

# now merging this xwalk (which has the Census GEOIDs) with the policing data
data_no_census <- left_join(x = data_no_census,
                            y = xwalk,
                            by = "city_name")

# checking random sample
data_no_census %>%
  select(year,city_name,GEOID.y) %>%
  sample_n(size = 10)

# renaming GEOID.y as GEOID
data_no_census %<>% dplyr::rename(GEOID = GEOID.y)

# turning GEOID into a character and adding leading 0
data_no_census$GEOID <- as.character(data_no_census$GEOID)
data_no_census$GEOID <- paste("0", data_no_census$GEOID, sep="")

# turning year into numeric
data_no_census$year <- as.numeric(data_no_census$year)

# now joining nearly full dataset minus segregation indices
data_no_seg <- left_join(x = data_no_census,
                         y = combined_data,
                         by = c("year", "GEOID"))

# last step is adding population density variable (one of my control variables)
load("data/pop_density.RData")

# rename fips as GEOID
pop_density %<>% dplyr::rename(GEOID = fips)

# turning GEOID into a character and adding leading 0
pop_density$GEOID <- as.character(pop_density$GEOID)
pop_density$GEOID <- paste("0", pop_density$GEOID, sep="")

# now joining full dataset minus segregation indices
data_no_seg <- left_join(x = data_no_seg,
                         y = pop_density,
                         by = c("year", "GEOID"))

# dropping place name and total pop
data_no_seg <- data_no_seg[-c(117,118)]

```

## 1.3 Interpolating Census Data

``` {r}

# interpolating all data with "rent" in the column name
interp <- data_no_seg %>%
  group_by(GEOID) %>%
  mutate_at(vars(matches("rent")), list(interp=na.approx), na.rm=FALSE)

# testing if that worked
#before interpolation
ggplot(subset(interp,city_name %in% c("Adelanto", "Antioch", "Los Angeles", "Lancaster", "Victorville", "Oakland", "San Francisco")), aes(x=year, y=prenter, color = city_name)) +
  geom_point() 

# after interpolation
ggplot(subset(interp,city_name %in% c("Adelanto", "Antioch", "Los Angeles", "Lancaster", "Victorville", "Oakland", "San Francisco")), aes(x=year, y=prenter_interp, color = city_name)) +
  geom_point() 
# it worked!

```

This plot shows one of the key trends of the last twenty years that I am interested in learning more about: Black renters have decreased in core cities like Oakland, SF, and LA, but are increasing in select exurbs like Lancaster and Victorville.
```{r}
ggplot(subset(interp,city_name %in% c("Los Angeles", "Lancaster", "Antioch", "Victorville", "Oakland", "San Francisco")), aes(x=year, y=pblkrenter_interp, color = city_name)) +
  geom_point() +
  geom_smooth(method = "lm") +
  xlim(2000,2020) +
  ggtitle("Percent of Renters Who Are Black") +
  labs(y="Percent Black Renters", x = "Year", color = "Select Cities") +
  theme_economist() + 
  scale_colour_economist() 

ggsave(filename="plots/pblkrenters.pdf")

```

``` {r}

# interpolating all data with "pov" in the column name (for poverty vars)
interp %<>%
  group_by(GEOID) %>%
  mutate_at(vars(matches("pov")), list(interp=na.approx), na.rm=FALSE)

# interpolating all data with "nh" in the column name (for nonhispanic vars)
interp %<>%
  group_by(GEOID) %>%
  mutate_at(vars(matches("nh")), list(interp=na.approx), na.rm=FALSE)

# interpolating all data with "his" in the column name (for nonhispanic vars)
interp %<>%
  group_by(GEOID) %>%
  mutate_at(vars(matches("his")), list(interp=na.approx), na.rm=FALSE)

# interpolating all data with "immig" in the column name (for immigrant vars)
interp %<>%
  group_by(GEOID) %>%
  mutate_at(vars(matches("immig")), list(interp=na.approx), na.rm=FALSE)

# interpolating all data with "pop" in the column name (for total population and population density vars)
interp %<>%
  group_by(GEOID) %>%
  mutate_at(vars(matches("pop")), list(interp=na.approx), na.rm=FALSE)

# interpolating all data with "blk_m" in the column name (for young Black men vars)
interp %<>%
  group_by(GEOID) %>%
  mutate_at(vars(matches("blk_m")), list(interp=na.approx), na.rm=FALSE)

```

## 1.4 Saving Data

Now I would like to save out a nearly completed dataset with just years ranging from 1998 to 2018.

``` {r}

# creating new dataframe for 1998-2018
all_data <- interp %>%
  subset(year>1997)

# save dataset
save(all_data,
     file = "data/all_data.RData")

```
